@@ -12,17 +12,17 @@ android {
     }
 
     compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_17
-        targetCompatibility = JavaVersion.VERSION_17
+        sourceCompatibility = JavaVersion.VERSION_21
+        targetCompatibility = JavaVersion.VERSION_21
     }
 
 }
 
 dependencies {
-    api("androidx.annotation:annotation:1.7.1")
+    api("androidx.annotation:annotation:1.8.1")
     implementation(project(":appcompat:appcompat"))
     api(project(":core:core"))
-    api("androidx.dynamicanimation:dynamicanimation:1.0.0")
+    api(libs.androidx.dynamicanimation)
     api(project(":viewpager:viewpager"))
     implementation libs.constraintLayout
 }
