@@ -12,14 +12,14 @@ android {
     }
 
     compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_17
-        targetCompatibility = JavaVersion.VERSION_17
+        sourceCompatibility = JavaVersion.VERSION_21
+        targetCompatibility = JavaVersion.VERSION_21
     }
 }
 
 dependencies {
     implementation(project(":appcompat:appcompat"))
-    implementation("sesl.com.google.android.material:material:1.12.0+1.0.23-sesl6+rev0")
+    implementation("sesl.com.google.android.material:material:1.12.0+1.0.22-sesl7+rev0")
     api(project(":core:core"))
     api(project(":fragment:fragment"))
 }
